Checkpoint - move gordo specific funcs into module
